requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
astunparse==1.6.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
crontab==1.0.1
cryptography==42.0.5
distro==1.9.0
dnspython==2.6.1
dotml==0.1.9
ecdsa==0.18.0
email_validator==2.1.1
fastapi==0.109.2
fastapi-cli==0.0.4
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
freezegun==1.5.0
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.65.1
h11==0.14.0
h5py==3.11.0
haystack-ai==2.3.0
haystack-experimental==0.1.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.5
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.3
json5==0.9.25
keras==3.4.1
lazy-imports==0.3.1
libclang==18.1.1
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.0
monotonic==1.6
more-itertools==10.3.0
multidict==6.0.5
mysql-connector-python==9.0.0
namex==0.0.8
networkx==3.3
nodeenv==1.8.0
numpy==1.26.4
openai==1.35.14
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.3
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.1
passlib==1.7.4
posthog==3.5.0
prisma==0.13.1
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.1
pyodbc==5.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
regex==2024.5.15
requests==2.31.0
rich==13.7.1
rq==1.16.2
rq-dashboard-fast==0.3.8
rq-scheduler==0.13.1
rsa==4.9
safetensors==0.4.3
schedule==1.2.2
setuptools==69.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.31
starlette==0.36.3
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.4.0
tf_keras==2.17.0
tokenizers==0.19.1
tomlkit==0.12.4
tqdm==4.66.4
transformers==4.42.4
trio==0.25.0
trio-websocket==0.11.1
typer==0.12.3
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
